requirements.txt
aiofiles==24.1.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
brotli==1.2.0
cachetools==5.5.2
certifi==2025.11.12
chardet==3.0.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.124.4
ffmpy==1.0.0
filelock==3.20.0
fpdf==1.7.2
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
googletrans==4.0.0rc1
gradio==6.1.0
gradio_client==2.0.1
groovy==0.1.2
h11==0.16.0
h2==3.2.0
hf-xet==1.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.13.0
numpy==1.26.4
ollama==0.6.1
opencv-python==4.9.0.80
orjson==3.11.5
packaging==23.2
pandas==2.3.3
pillow==10.2.0
protobuf==4.25.8
pyarrow==22.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.30.0
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.50.0
streamlit==1.32.2
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.3
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
watchdog==6.0.0
